requirements.txt
django>=5.2
djangorestframework>=3.17
django-cors-headers>=4.9
whitenoise>=6.12
gunicorn>=25.0
